Run qemu in docker to improve isolation
The included Dockerfile builds a minimal, statically linked copy of qemu
and puts it in an otherwise empty docker container. The build runner now
runs its VMs inside of that container. Now there are 3 layers of
isolation between build guests and the rest of sr.ht:

1. KVM isolation
2. LXC isolation
3. Build runners are physically separate from the rest of sr.ht and have
   minimal access to sr.ht resources

Still TODO in isolation improvements is the build runner API, which
would replace the limited SQL access they have today.