Refactor Arch Linux genimg and control script
This refactors the Arch Linux genimg to use qemu-nbd instead of
libguestfs. It also uses a static IP now, and properly deals with the
fact that we don't want to update linux/mkinitcpio.

The way variants specify their extra boot criteria has been refactored
to better support more images in the future.

Using docker is now optional, which will be useful for sanity checks.