images/archlinux: do not run unneeded pacman hooks
The hook installed by the kernel for the purpose of running mkinitcpio
is not needed during CI runs. The package itself is in IgnorePkg to
prevent useless updates, however, currently, it is triggered regardless by
updates which distribute files in /usr/lib/initcpio. For example,
systemd.

Mask the hook file so that it does not get run under any circumstances.