Fix generation of /boot/vmlinuz* in F31+ images
The kernel-install script (provided by systemd-udev) does not install
kernel images in /boot if the systemd machine ID is not defined *AND* if
grubby is not installed.

I'd like to get rid of grubby since it depends on grub2-tools (does not
make sense since we use syslinux) but I believe it is good enough for
now, as it would require more digging that I can do right now.