OpenBSD 6.7
This modifies genimg to be able to build a 6.7 image.

Due to OpenBSD ABI issues I've changed the build manifest not to use the
openbsd/latest image as each image should preferably be built using the
same version of OpenBSD. When building from 6.6 the script produces a
slightly hacky "bootstrap image" which should then be used to
immediately re-build a proper image.

To bootstrap, you should be able to take the updated build.yaml and run
it using a 6.6 image manually.

I've tested building the bootstrap 6.7[0] and 6.6 with syspatch[1] so
I'm reasonably confident that 6.7 will also build fine once the
bootstrap image exists.

[0] https://builds.sr.ht/~oranenj/job/215911
[1] https://builds.sr.ht/~oranenj/job/215924