Disable APT::Install-Recommends in Debian images
Unlike the desktop and the manual host setup cases, a build environment
should err on the side of installing less software. Not just to optimize
build times, but also to help keep builds repeatable (so that strict
build dependencies have to be explicitly spelled out) and reliable (so
that a stray optional package doesn't break the build or subtly change
build results).