images/freebsd: use git to grab ports
Prior to the commit, we used snaphosts. Snapshots are automatically
generated every week. Their URL is quite unstable, and a release's
snapshot directory may be deleted without prior notice.

Since we just want to grab a snapshot of the ports tree, directly
download it from the source repository. For now we use GitHub as the
upstream, but we should be able to use freebsd.org's Git repo once
they're done with their Git migration.

Download a shallow clone so that our users can easily refresh the ports
collection if they need it. The Git data accounts for 82MiB at the time
of writing.