Add an index for owner_id to jobs table
The api.jobs_GET endpoint currently has very poor performance
(~8s latency) because it has to do a full table scan. The same query is
performed by jobs.index but it appears that the performance measurement
there is drowned out by not logged in users opening the page, which
skips the query. Adding an index on owner_id should change the full
table scan to an index lookup, improving performance.