api: drop mutation { cancelGroup }
Cancellation is weird. We can only really cancel jobs which are already
running, and in a job group, it's likely that there are a mix of jobs in
the running and queued states. We will probably support cancelling
queued or pending jobs after the coming builds overhaul, but in the
meantime I don't want to encode any of the present weirdness into an API
which would later become stable.