Revert "images/archlinux: skip compression for makepkg"
This reverts commit 73b24a385c95ef2df0d18bff1545980e9fc6f35c.

This affects packages which are built and uploaded in a builds.sr.ht
job, not just packages which are built as a side effect of installing
AUR packages in the job setup.

In commit e6d7370899cfc125ff7db7c61a31c52fa9049b0a the compression for
installing AUR packages in the job setup, was already disabled, using
the same method as this commit performed but scoped only to `yay -Syu`.

To quote the commit message:

"This patch has no effect on any makepkg run by the user in its build
script after the install() step ends."

which I think is a reasonable approach to take.

Signed-off-by: Eli Schwartz <eschwartz@archlinux.org>