images/debian,ubuntu: Replace deprecated apt-key
Since the use of apt-key has been deprecated ([1]), the Debian and
Ubuntu `add_repository` functions have had their invocations replaced
with two calls to `gpg`:

1. Given a key-id, fetch the key from keyserver.ubuntu.com and store it
   in a new trust database.
2. Export that same key from the trust database to a file in
   `/etc/apt/trusted.gpg.d/`, as per the deprecation instructions.

Ideally this could all be done in a single step, but I'm not aware of a
way to do that.

Once this is merged, the builds.sr.ht compatibility docs should be
updated to remove the mention of `apt-key` ([2]).

[1]: https://www.debian.org/releases/bullseye/amd64/release-notes/ch-information.en.html#deprecated-components
[2]: https://git.sr.ht/~sircmpwn/sr.ht-docs/tree/ac19223120f57fbbd88ec8c784250b0343a6405d/item/builds.sr.ht/compatibility.md?view-source#L413