when failing to clone a repository, include the used protocol
It fits decently well in the error message, and can help people catch
assumptions such as "I tried to clone a mercurial repository without
hg+", which resulted in builds.sr.ht thinking it was git, and attempting
to use `git clone`.

The error messages from git/hg themselves are not super obvious.

For `git clone` it prints:
```
fatal: repository 'https://hg.sr.ht/~username/reponame/' not found
```

For `hg clone` it prints:
```
abort: HTTP Error 404: NOT FOUND
```

Neither one makes it obvious that you attempted to clone with the wrong
VCS program, and mercurial also doesn't make it obvious if you attempted
to clone with the wrong url, but it's particularly non-obvious to debug
the case of accidentally using git.