Fix qemu image build on latest Debian unstable
In the latest Debian unstable (bookworm), `/bin` and `/lib` are symlinks
to `/usr/bin` and `/usr/lib` respectively. Hence, even though we
configure with e.g. `bindir=/bin`, the files get installed to
`/usr/bin` (same for the libs).

Docker COPY however does not resolve symlinks, so the actual path has to
be used. To avoid confusion, the arguments to `configure` have also been
updated (essentially a no-op on that image).