api/artifacts: Use SelectAll and ScanAll
While working on hut I found a problem with my last patch concerning the
pruned URL feature [1]. The current implementation only works if the
GraphQL query contains the "created" timestamp, otherwise the "url"
field will always be null. That is caused by the zero value for time
which is used when created is not populated by the database query.

[1]: https://lists.sr.ht/~sircmpwn/sr.ht-dev/patches/39275