API: fetch user context in loader implementation
Fetching it outside of this function will cause it to attempt to fetch
the auth context when the middleware is stood up, which occurs
unconditionally on all requests -- including unauthenticated ones, such
as /query/metrics. Moving it into the function body causes it to only
occur once we actually attempt to fetch jobs.