Track number of builds currently running and export number of build slots
Duration histogram tracks the number of builds finished, and having
the number of builds running, number of builds started is trivial to
calculate, while having a gauge for currently running jobs makes it
easier to find current utilization. Additionally, a status label for
duration metric is added, to keep track of the distribution of different
build statuses that was kept by the builds finished metric.