images/control: Clarify error messages
Changes the error message for trying to run a non-existent architecture
to mention that it's specifically the architecture that's missing and
not the specific image (which is checked at the very start of the file).

The error message for a non-existent image is clarified a bit too, to
make it more clear what's wrong in the manifest, instead of how the
runner is detecting that.