api/graph: check SECRETS:RO in submit(secrets: true) mutation
Submitting a build with secrets enabled grants access to secrets.
Let's reflect this in the token scope requirements.

In order to not break builds with no secrets argument specified,
make the default value a bit smarter: enable secrets if at least
one is specified in the manifest and the SECRETS:RO grant is
available.