Use correct internal URL for fetching logs
In a setup where internal traffic is meant to stay internal (i.e. the
`internal-origin` config is used) fetching logs is currently broken. The
logs are served by the api component, but as is the request will be sent
to the Python frontend.

Fix this by using the value of `api-origin`, just falling back to
`get_origin` if that is not set. This is the same technique already
employed in the Python GraphQL client.

Signed-off-by: Conrad Hoffmann <ch@bitfehler.net>